fix: add search to settings
