+++ apps/studio/src/templates/layouts/AdminCmsSidebarLayout.tsx
@@ -1,9 +1,6 @@
 import type { PropsWithChildren } from "react"
-import Image from "next/image"
-import NextLink from "next/link"
 import { useRouter } from "next/router"
-import { Flex, Grid, GridItem, Text } from "@chakra-ui/react"
-import { AvatarMenu, IconButton, Menu } from "@opengovsg/design-system-react"
+import { Flex } from "@chakra-ui/react"
 import { BiCog, BiFolder, BiHelpCircle, BiLogOut } from "react-icons/bi"
 import { z } from "zod"
 
@@ -11,70 +8,11 @@ import type { CmsSidebarItem } from "~/components/CmsSidebar/CmsSidebarItems"
 import { EnforceLoginStatePageWrapper } from "~/components/AuthWrappers"
 import { CmsContainer, CmsSidebar } from "~/components/CmsSidebar"
 import { LayoutHead } from "~/components/LayoutHead"
-import { Searchbar, useSearchStyle } from "~/components/Searchbar"
+import { SearchableHeader } from "~/components/SearchableHeader"
 import { DirectorySidebar } from "~/features/dashboard/components/DirectorySidebar"
 import { useMe } from "~/features/me/api"
 import { useQueryParse } from "~/hooks/useQueryParse"
-import { DASHBOARD } from "~/lib/routes"
 import { type GetLayout } from "~/lib/types"
-import { trpc } from "~/utils/trpc"
-
-interface SearchableHeaderProps {
-  siteId: string
-}
-const SearchableHeader = ({ siteId }: SearchableHeaderProps) => {
-  const { me, logout } = useMe()
-  const [{ name }] = trpc.site.getSiteName.useSuspenseQuery({
-    siteId: Number(siteId),
-  })
-  const { minWidth, maxWidth } = useSearchStyle()
-
-  return (
-    <Grid
-      gridTemplateColumns={`1fr ${maxWidth} 1fr`}
-      py={{ base: 0, md: "0.5rem" }}
-      px={{ base: 0, md: "0.5rem" }}
-      background="white"
-    >
-      <Flex alignItems="center" as={GridItem}>
-        <IconButton
-          mr="0.5rem"
-          as={NextLink}
-          href={DASHBOARD}
-          variant="clear"
-          aria-label="Back to dashboard"
-          icon={
-            <Image
-              src="/assets/isomer-logo-color.svg"
-              height={24}
-              width={22}
-              alt="Back to dashboard"
-              priority
-            />
-          }
-        />
-        <Text textStyle="subhead-2" noOfLines={1}>
-          {name}
-        </Text>
-      </Flex>
-      {/* NOTE: We are doing this because the searchbar has to be horizontally centered within the Flex */}
-      <GridItem minW={minWidth} maxW={maxWidth}>
-        <Searchbar siteId={siteId} />
-      </GridItem>
-
-      <GridItem justifyContent="flex-end" display="flex">
-        <AvatarMenu
-          name={me.name}
-          variant="subtle"
-          bg="base.canvas.brand-subtle"
-          menuListProps={{ maxWidth: "19rem" }}
-        >
-          <Menu.Item onClick={() => logout()}>Sign out</Menu.Item>
-        </AvatarMenu>
-      </GridItem>
-    </Grid>
-  )
-}
 
 export const AdminCmsSearchableLayout: GetLayout = (page) => {
   return (
